requirements.txt
einops
requests
pytest
tqdm
datasets
transformers
sentencepiece
notebook
ipykernel
nbformat>=4.2.0
ipywidgets==7.7.1
jupyterlab-widgets==1.1.1
protobuf==3.20.0
jupyter
matplotlib
numpy-stl
pandas
scikit-learn
wandb==0.13.10
plotly
gdown
jaxtyping
streamlit
streamlit-analytics
streamlit-antd-components
torchinfo
frozendict
openai
transformer_lens
git+https://github.com/callummcdougall/CircuitsVis.git#subdirectory=python
git+https://github.com/callummcdougall/eindex.git